requirements.txt
asgiref==3.2.3
Django>=2.1.3,<2.2.0
djangorestframework>=3.9.0,<3.10.0
django-cors-headers
psycopg2>=2.7.5,<2.8.0
Pillow>=5.3.0,<5.4.0
django-pyodbc-azure==2.1.0.0
pyodbc==4.0.28
pytz==2019.3
sqlparse==0.3.0
